Time: 12 ms (77.49%), Space: 9.9 MB (97.92%) - LeetHub

+++ 0518-coin-change-ii/0518-coin-change-ii.cpp
@@ -1,17 +1,17 @@
 class Solution {
 public:
     int change(int amount, vector<int>& coins) {
-        vector<int> dp(amount + 1);
+        vector<unsigned int> dp(amount + 1);
         dp.back() = 1;
         for(int r = coins.size() - 1; r >= 0; r--){
             for(int c = amount; c >= 0; c--){
                 int coin = coins[r];
-                int right = 0;
+                unsigned int right = 0;
                 if(c + coin <= amount) {
                     right = dp[c + coin];
                 }
-                int below = dp[c];
-                long long ans = (long long)right + below;
+                unsigned int below = dp[c];
+                unsigned int ans = right + below;
                 if(ans > INT_MAX) {
                     ans = 0;
                 }
